Se realiza la totalidad de los diseños requeridos.

+++ RacePuntos/Controllers/personasController.cs
@@ -15,26 +15,41 @@ namespace RacePuntos.Controllers {
 
 		// GET: personas
 		public async Task<ActionResult> Index() {
-			var personas = db.personas.Include(p => p.cargos);
-			return View(await personas.ToListAsync());
+			if(Session["USUARIO_LOGUEADO"] != null) {
+				var personas = db.personas.Include(p => p.cargos);
+				return View(await personas.ToListAsync());
+			} else {
+				Response.Redirect("/Personas/Login");
+				return null;
+			}
 		}
 
 		// GET: personas/Details/5
 		public async Task<ActionResult> Details(string id) {
-			if(id == null) {
-				return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+			if(Session["USUARIO_LOGUEADO"] != null) {
+				if(id == null) {
+					return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+				}
+				personas personas = await db.personas.FindAsync(id);
+				if(personas == null) {
+					return HttpNotFound();
+				}
+				return View(personas);
+			} else {
+				Response.Redirect("/Personas/Login");
+				return null;
 			}
-			personas personas = await db.personas.FindAsync(id);
-			if(personas == null) {
-				return HttpNotFound();
-			}
-			return View(personas);
 		}
 
 		// GET: personas/Create
 		public ActionResult Create() {
-			ViewBag.cargo = new SelectList(db.cargos, "id_cargo", "nombre_cargo");
-			return View();
+			if(Session["USUARIO_LOGUEADO"] != null) {
+				ViewBag.cargo = new SelectList(db.cargos, "id_cargo", "nombre_cargo");
+				return View();
+			} else {
+				Response.Redirect("/Personas/Login");
+				return null;
+			}
 		}
 
 		// POST: personas/Create
@@ -43,27 +58,37 @@ namespace RacePuntos.Controllers {
 		[HttpPost]
 		[ValidateAntiForgeryToken]
 		public async Task<ActionResult> Create([Bind(Include = "tipo_documento,documento,contrasena,rol,cargo,id_usuario_creacion,nombres,apellidos,fecha_nacimiento,direccion,numero_celular,correoElectronico")] personas personas) {
-			if(ModelState.IsValid) {
-				db.personas.Add(personas);
-				await db.SaveChangesAsync();
-				return RedirectToAction("Index");
-			}
+			if(Session["USUARIO_LOGUEADO"] != null) {
+				if(ModelState.IsValid) {
+					db.personas.Add(personas);
+					await db.SaveChangesAsync();
+					return RedirectToAction("Index");
+				}
 
-			ViewBag.cargo = new SelectList(db.cargos, "id_cargo", "nombre_cargo", personas.cargo);
-			return View(personas);
+				ViewBag.cargo = new SelectList(db.cargos, "id_cargo", "nombre_cargo", personas.cargo);
+				return View(personas);
+			} else {
+				Response.Redirect("/Personas/Login");
+				return null;
+			}
 		}
 
 		// GET: personas/Edit/5
 		public async Task<ActionResult> Edit(string id) {
-			if(id == null) {
-				return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+			if(Session["USUARIO_LOGUEADO"] != null) {
+				if(id == null) {
+					return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+				}
+				personas personas = await db.personas.FindAsync(id);
+				if(personas == null) {
+					return HttpNotFound();
+				}
+				ViewBag.cargo = new SelectList(db.cargos, "id_cargo", "nombre_cargo", personas.cargo);
+				return View(personas);
+			} else {
+				Response.Redirect("/Personas/Login");
+				return null;
 			}
-			personas personas = await db.personas.FindAsync(id);
-			if(personas == null) {
-				return HttpNotFound();
-			}
-			ViewBag.cargo = new SelectList(db.cargos, "id_cargo", "nombre_cargo", personas.cargo);
-			return View(personas);
 		}
 
 		// POST: personas/Edit/5
@@ -72,35 +97,60 @@ namespace RacePuntos.Controllers {
 		[HttpPost]
 		[ValidateAntiForgeryToken]
 		public async Task<ActionResult> Edit([Bind(Include = "tipo_documento,documento,contrasena,rol,cargo,id_usuario_creacion,nombres,apellidos,fecha_nacimiento,direccion,numero_celular,correoElectronico")] personas personas) {
-			if(ModelState.IsValid) {
-				db.Entry(personas).State = EntityState.Modified;
-				await db.SaveChangesAsync();
-				return RedirectToAction("Index");
+			if(Session["USUARIO_LOGUEADO"] != null) {
+				if(ModelState.IsValid) {
+					db.Entry(personas).State = EntityState.Modified;
+					await db.SaveChangesAsync();
+					return RedirectToAction("Index");
+				}
+				ViewBag.cargo = new SelectList(db.cargos, "id_cargo", "nombre_cargo", personas.cargo);
+				return View(personas);
+			} else {
+				Response.Redirect("/Personas/Login");
+				return null;
 			}
-			ViewBag.cargo = new SelectList(db.cargos, "id_cargo", "nombre_cargo", personas.cargo);
-			return View(personas);
 		}
 
 		// GET: personas/Delete/5
 		public async Task<ActionResult> Delete(string id) {
-			if(id == null) {
-				return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+			if(Session["USUARIO_LOGUEADO"] != null) {
+				if(id == null) {
+					return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+				}
+				personas personas = await db.personas.FindAsync(id);
+				if(personas == null) {
+					return HttpNotFound();
+				}
+				return View(personas);
+			} else {
+				Response.Redirect("/Personas/Login");
+				return null;
 			}
-			personas personas = await db.personas.FindAsync(id);
-			if(personas == null) {
-				return HttpNotFound();
-			}
-			return View(personas);
 		}
 
 		// POST: personas/Delete/5
 		[HttpPost, ActionName("Delete")]
 		[ValidateAntiForgeryToken]
 		public async Task<ActionResult> DeleteConfirmed(string id) {
-			personas personas = await db.personas.FindAsync(id);
-			db.personas.Remove(personas);
-			await db.SaveChangesAsync();
-			return RedirectToAction("Index");
+			if(Session["USUARIO_LOGUEADO"] != null) {
+				personas personas = await db.personas.FindAsync(id);
+				db.personas.Remove(personas);
+				await db.SaveChangesAsync();
+				return RedirectToAction("Index");
+			} else {
+				Response.Redirect("/Personas/Login");
+				return null;
+			}
+		}
+
+		// GET: personas/servicios
+		public async Task<ActionResult> Servicios() {
+			if(Session["USUARIO_LOGUEADO"] != null) {
+				return View();
+			} else {
+				Response.Redirect("/Personas/Login");
+				return null;
+			}
 		}
 
 		//
@@ -125,15 +175,15 @@ namespace RacePuntos.Controllers {
 
 				var ValidateLogin = db.logeo_persona(documento.ToString(), contrasena.ToString()).Count();
 
-				
+
 				if(ValidateLogin > 0) {
 					personas User = new personas {
 						documento = documento,
 						nombres = db.personas.Where(c => c.documento == documento).First().nombres,
-						apellidos = db.personas.Where(c=>c.documento == documento).First().apellidos
+						apellidos = db.personas.Where(c => c.documento == documento).First().apellidos
 					};
 					Session["USUARIO_LOGUEADO"] = User;
-					
+
 					Session["DOCUMENTO"] = db.personas.Where(c => c.documento == documento).First().documento;
 					Session["NOMBRES"] = db.personas.Where(c => c.documento == documento).First().nombres;
 					Session["APELLIDOS"] = db.personas.Where(c => c.documento == documento).First().apellidos;
@@ -141,6 +191,9 @@ namespace RacePuntos.Controllers {
 					Session["CARGO"] = db.personas.Where(c => c.documento == documento).First().cargo;
 
 					Response.Redirect("/Home/Index");
+					return null;
+				} else {
+					Response.Redirect("~/Personas/Login");
 					return null;
 				}
 			} catch(Exception ex) {
